benchmarks: use heavier input in benchmarking (1-100MB)
