+++ .gitignore
@@ -1,3 +1,3 @@
-test/*.json
-test/test
+benchmarks/*.json
+benchmarks/test
 *.pprof
